requirements.txt
beautifulsoup4==4.12.3
cdsapi==0.7.1
chromadb==0.4.24
colorama==0.4.6
filelock==3.15.4
fiona==1.9.6
geopandas==1.0.1
groq==0.9.0
langchain==0.2.14
langchain_community==0.2.12
langchain_core==0.2.35
langchain_experimental==0.0.64
langchain_groq==0.1.9
langchain_openai==0.1.22
matplotlib==3.8.3
msgpack_python==0.5.6
netCDF4==1.7.1.post2
numpy>=1.22.5
openai==1.42.0
opencv_python==4.10.0.84
pandas==2.2.2
panflute==2.3.1
Pillow==10.4.0
psutil==5.9.8
pyclipper==1.3.0.post5
PyOpenGL==3.1.7
PyQt5==5.15.11
PyQt6_sip==13.8
pytest==8.3.2
python-dotenv==1.0.1
PyYAML==6.0.2
pyzmq==25.1.2
Requests==2.32.3
rich==13.8.0
rtree==1.3.0
scipy==1.14.1
seaborn==0.13.2
setuptools==56.0.0
Shapely==2.0.6
streamlit==1.32.2
textual==0.78.0
watchdog==4.0.0
#bluesky gui requirements
PyQt6
PyQt6-WebEngine
pygame
pyopengl
msgpack
zmq
bluesky-simdata
bluesky-guidata
